# benchmark single list vs iterator for active fps (DNM)