--- a/x/finality/keeper/power_table.go
+++ b/x/finality/keeper/power_table.go
@@ -142,10 +142,63 @@ func (k Keeper) newVotingPowerStore(ctx context.Context) prefix.Store {
 	return prefix.NewStore(storeAdapter, types.VotingPowerAsListKey)
 }
 
+type ActiveFp struct {
+	FPBTCPK     []byte
+	VotingPower uint64
+}
+
+func (afp *ActiveFp) Marshal() []byte {
+	var buf bytes.Buffer
+	buf.Write(afp.FPBTCPK)
+	buf.Write(sdk.Uint64ToBigEndian(afp.VotingPower))
+	return buf.Bytes()
+}
+
+func (afp *ActiveFp) Unmarshal(data []byte) error {
+	buf := bytes.NewBuffer(data)
+	afp.FPBTCPK = buf.Next(32)
+	afp.VotingPower = sdk.BigEndianToUint64(buf.Next(8))
+	return nil
+}
+
+type ActiveFpList struct {
+	Fps []*ActiveFp
+}
+
+func (afl *ActiveFpList) Marshal() []byte {
+	var buf bytes.Buffer
+	for _, fp := range afl.Fps {
+		buf.Write(fp.Marshal())
+	}
+	return buf.Bytes()
+}
+
+func (afl *ActiveFpList) Unmarshal(data []byte) error {
+	fps := []*ActiveFp{}
+	buf := bytes.NewBuffer(data)
+	for buf.Len() > 0 {
+		fp := &ActiveFp{}
+		if err := fp.Unmarshal(buf.Next(40)); err != nil {
+			return err
+		}
+		fps = append(fps, fp)
+	}
+	afl.Fps = fps
+	return nil
+}
+
 func (k Keeper) SetVotingPowerAsList(ctx context.Context, height uint64, activeFPs []*types.ActiveFinalityProvider) {
 	store := k.newVotingPowerStore(ctx)
 	activeList := types.ActiveFinalityProvidersList{Fps: activeFPs}
 	activeListBytes := k.cdc.MustMarshal(&activeList)
+
+	store.Set(sdk.Uint64ToBigEndian(height), activeListBytes)
+}
+
+func (k Keeper) SetVotingPowerAsListNew1(ctx context.Context, height uint64, activeFPs []*ActiveFp) {
+	store := k.newVotingPowerStore(ctx)
+	activeList := ActiveFpList{Fps: activeFPs}
+	activeListBytes := activeList.Marshal()
 
 	store.Set(sdk.Uint64ToBigEndian(height), activeListBytes)
 }
@@ -182,6 +235,30 @@ func (k Keeper) GetVotingPowerNew(ctx context.Context, fpBTCPK []byte, height ui
 
 	for _, fp := range activeList.Fps {
 		if bytes.Equal(fp.FpBtcPk.MustMarshal(), fpBTCPK) {
+			return fp.VotingPower
+		}
+	}
+
+	return 0
+}
+
+func (k Keeper) GetVotingPowerNew1(ctx context.Context, fpBTCPK []byte, height uint64) uint64 {
+	store := k.newVotingPowerStore(ctx)
+	activeListBytes := store.Get(sdk.Uint64ToBigEndian(height))
+
+	if activeListBytes == nil {
+		return 0
+	}
+
+	buf := bytes.NewBuffer(activeListBytes)
+
+	for buf.Len() > 0 {
+		fp := &ActiveFp{}
+		if err := fp.Unmarshal(buf.Next(40)); err != nil {
+			return 0
+		}
+
+		if bytes.Equal(fp.FPBTCPK, fpBTCPK) {
 			return fp.VotingPower
 		}
 	}